+++ src/components/NavBar/NavBar.tsx
@@ -5,7 +5,7 @@ import { Avatar, Button, List, Stack, Toolbar } from "@mui/material";
 import { Drawer } from "@mui/material";
 import AccountMenu from "./AccountMenu";
 import NavItem from "./NavItem";
-import PostButtonModal from "../Misc/PostButtonModal";
+import PostButtonModal from "./PostButtonModal";
 import { useState } from "react";
 import ComposePost from "../Posts/ComposePost";
