+++ src/components/Posts/ExpandedPostReplies.tsx
@@ -4,25 +4,31 @@ import axios from "axios";
 import { useAppDispatch, useAppSelector } from "../../state/hooks";
 import { Post, setPosts } from "../../state/slices/postsSlice";
 
-const ProfileReplies = () => {
+const ExpandedPostReplies = () => {
   const posts = useAppSelector((state) => state.posts);
   const user = useAppSelector((state) => state.user);
+  const post = useAppSelector((state) => state.post);
   const dispatch = useAppDispatch();
 
   useEffect(() => {
     const fetchPosts = async () => {
-      const result = await axios.get(
-        "http://localhost:3001/api/profile/getOwnReplies",
-        {
-          params: {
-            userId: user.userId,
-          },
-        }
-      );
-      dispatch(setPosts(result.data as Post[]));
+      try {
+        const resultReplies = await axios.get(
+          "http://localhost:3001/api/posts/fetchReplies",
+          {
+            params: {
+              userId: user.userId,
+              postId: post.postId,
+            },
+          }
+        );
+        dispatch(setPosts(resultReplies.data as Post[]));
+      } catch (e) {
+        console.log(e.message);
+      }
     };
     fetchPosts();
-  }, [dispatch, user]);
+  }, [dispatch, user, post]);
 
   return (
     <>
@@ -33,4 +39,4 @@ const ProfileReplies = () => {
   );
 };
 
-export default ProfileReplies;
+export default ExpandedPostReplies;
