+++ src/pages/Profile.tsx
@@ -3,11 +3,11 @@ import { useEffect } from "react";
 import KeyboardBackspaceIcon from "@mui/icons-material/KeyboardBackspace";
 import { Button, Typography, Tabs, Tab, Avatar, Box } from "@mui/material";
 import EditIcon from "@mui/icons-material/Edit";
-import ProfileTweets from "../components/Posts/ProfileTweets";
+import ProfileTweets from "../components/Profile/ProfileTweets";
 import axios from "axios";
 import { useAppSelector } from "../state/hooks";
-import ProfileReplies from "../components/Posts/ProfileReplies";
-import ProfileLikes from "../components/Posts/ProfileLikes";
+import ProfileReplies from "../components/Profile/ProfileReplies";
+import ProfileLikes from "../components/Profile/ProfileLikes";
 import CalendarMonthIcon from "@mui/icons-material/CalendarMonth";
 import { useNavigate } from "react-router-dom";
 import IconButton from "@mui/material/IconButton/IconButton";
